requirements.txt
pymongo
pyrogram>=2.0.0
pyromod==1.5.0
python-dotenv
TgCrypto
